Use process_handler to clean up stack and make jmp
Don't use the resume_process function or either of the _resume_process assembly
interrupt handler routines. They leave a bunch of junk on the stack and lead to
all sorts of nasty overflows.

Instead:
1. Save the state of the current process into the process struct in
	scheduler_wakeup
2. Schedule the execution of the next process
3. Write to 0x20 to indicate end of interrupt
4. enable paging from the proper page directory. Note that right now both
processes share a single page directory! This is to reduce the places where
things can fail and surprise me.
5. Set the cpu state struct regs on the stack to be the one from the process we
are about to execute.
6. Return back into the process_handler
7. pop the segment descriptor
8. restore kernel segment descriptors
9. pop all registers
10. Remove error codes from stack (pushed earlier on lines 51 & 52)
11. Set interrupts. Takes effect after the next instruction completes.
12. Jump to the location of the instruction pointer in the regs struct (which
has actually been subtracted off the stack but not disturbed)

Note that something is still amiss. I may not be restoring all of my registers
properly. It seems that after a number of context switches something goes
berzerk and I get a low number in my regs struct's eip field.